Merged my Android Project with Jackies Android Project
I have created the tasker or to do list along with its database
processes including checking off of items. I have also made a dbHelper
to allow for easy control over the database in the near future. I have
created the Personalization of the mobile device to allow for the user
to choose sleep and wake up times for the device. And i got Github
working :dancer: